refactor (repository) : group drink-related files
