+++ src/main/java/kuit/project/beering/repository/drink/CustomDrinkRepository.java
@@ -1,4 +1,4 @@
-package kuit.project.beering.repository;
+package kuit.project.beering.repository.drink;
 
 import kuit.project.beering.dto.request.drink.DrinkSearchCondition;
 import kuit.project.beering.dto.response.drink.DrinkSearchResponse;
