+++ src/main/java/kuit/project/beering/repository/drink/DrinkRepository.java
@@ -1,4 +1,4 @@
-package kuit.project.beering.repository;
+package kuit.project.beering.repository.drink;
 
 import kuit.project.beering.domain.Drink;
 import org.springframework.data.jpa.repository.JpaRepository;
